Refactor moved handout processing to its own file
